Organize API and CSP (a.k.a. HSM/LASP) tests into folders/namespaces
