+++ tests/Agent/IntegrationTests/IntegrationTests/Api/AgentApiTests.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using NewRelic.Agent.IntegrationTestHelpers;
@@ -11,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.Api
 {
     [NetFrameworkTest]
     public class AgentApiTests : IClassFixture<RemoteServiceFixtures.AgentApiExecutor>
