+++ tests/Agent/IntegrationTests/IntegrationTests/Api/ApiCallsTests.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.Api
 {
     [NetFrameworkTest]
     public class ApiCallsTestsFW : ApiCallsTests<ConsoleDynamicMethodFixtureFW>
