+++ tests/Agent/IntegrationTests/IntegrationTests/CSP/AspNetCoreHSMEnabledWithCustomAttributesTests.cs
@@ -8,18 +8,18 @@ using System.Linq;
 using NewRelic.Agent.IntegrationTestHelpers;
 using NewRelic.Agent.IntegrationTestHelpers.Models;
 using NewRelic.Agent.IntegrationTests.RemoteServiceFixtures;
-using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.CSP
 {
-    [NetFrameworkTest]
-    public class HighSecurityAndCustomAttributes : IClassFixture<HSMCustomAttributesWebApi>
+    [NetCoreTest]
+    public class AspNetCoreHSMEnabledWithCustomAttributesTests : IClassFixture<HSMAspNetCoreWebApiCustomAttributesFixture>
     {
-        private readonly HSMCustomAttributesWebApi _fixture;
 
-        public HighSecurityAndCustomAttributes(HSMCustomAttributesWebApi fixture, ITestOutputHelper output)
+        private readonly HSMAspNetCoreWebApiCustomAttributesFixture _fixture;
+
+        public AspNetCoreHSMEnabledWithCustomAttributesTests(HSMAspNetCoreWebApiCustomAttributesFixture fixture, ITestOutputHelper output)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
@@ -29,14 +29,16 @@ namespace NewRelic.Agent.IntegrationTests
                     var configPath = fixture.DestinationNewRelicConfigFilePath;
                     var configModifier = new NewRelicConfigModifier(configPath);
                     configModifier.ForceTransactionTraces();
-                    configModifier.SetHighSecurityMode(true);
-                    configModifier.SetLogLevel("debug");
+
+                    CommonUtils.ModifyOrCreateXmlAttributeInNewRelicConfig(configPath, new[] { "configuration", "log" }, "level",
+                        "debug");
+                    CommonUtils.ModifyOrCreateXmlAttributeInNewRelicConfig(configPath, new[] { "configuration", "highSecurity" },
+                        "enabled", "true");
                 },
                 exerciseApplication: () =>
                 {
                     _fixture.Get();
-                    _fixture.GetCustomErrorAttributes();
-                    _fixture.AgentLog.WaitForLogLine(AgentLogFile.ErrorEventDataLogLineRegex, TimeSpan.FromMinutes(2));
+                    _fixture.AgentLog.WaitForLogLine(AgentLogFile.TransactionSampleLogLineRegex, TimeSpan.FromMinutes(2));
                 }
 
                 );
@@ -54,22 +56,7 @@ namespace NewRelic.Agent.IntegrationTests
 
             var transactionSample = _fixture.AgentLog.GetTransactionSamples().FirstOrDefault();
             Assert.NotNull(transactionSample);
-
             Assertions.TransactionTraceDoesNotHaveAttributes(unexpectedTransactionTraceAttributes, TransactionTraceAttributeType.User, transactionSample);
-
-            var errorTrace = _fixture.AgentLog.GetErrorTraces().FirstOrDefault();
-            var errorEventPayload = _fixture.AgentLog.GetErrorEvents().FirstOrDefault();
-
-            var unexpectedErrorCustomAttributes = new List<string>
-            {
-                "hey",
-                "faz"
-            };
-
-            NrAssert.Multiple(
-                () => Assertions.ErrorTraceDoesNotHaveAttributes(unexpectedErrorCustomAttributes, ErrorTraceAttributeType.Intrinsic, errorTrace),
-                () => Assertions.ErrorEventDoesNotHaveAttributes(unexpectedErrorCustomAttributes, EventAttributeType.User, errorEventPayload.Events[0])
-            );
         }
     }
 }
